+++ crates/eframe/src/native/epi_integration.rs
@@ -146,7 +146,7 @@ pub fn handle_app_output(
         fullscreen,
         drag_window,
         window_pos,
-        visible,
+        visible: _, // handled in post_present
         always_on_top,
     } = app_output;
 
@@ -181,10 +181,6 @@ pub fn handle_app_output(
 
     if drag_window {
         let _ = window.drag_window();
-    }
-
-    if let Some(visible) = visible {
-        window.set_visible(visible);
     }
 
     if let Some(always_on_top) = always_on_top {
@@ -240,7 +236,10 @@ impl EpiIntegration {
                 native_pixels_per_point: Some(native_pixels_per_point),
                 window_info: read_window_info(window, egui_ctx.pixels_per_point()),
             },
-            output: Default::default(),
+            output: epi::backend::AppOutput {
+                visible: Some(true),
+                ..Default::default()
+            },
             storage,
             #[cfg(feature = "glow")]
             gl,
@@ -325,6 +324,7 @@ impl EpiIntegration {
             if app_output.close {
                 self.close = app.on_close_event();
             }
+            self.frame.output.visible = app_output.visible; // this is handled by post_present
             handle_app_output(window, self.egui_ctx.pixels_per_point(), app_output);
         }
 
@@ -339,6 +339,12 @@ impl EpiIntegration {
         let window_size_px = [inner_size.width, inner_size.height];
 
         app.post_rendering(window_size_px, &self.frame);
+    }
+
+    pub fn post_present(&mut self, window: &winit::window::Window) {
+        if let Some(visible) = self.frame.output.visible.take() {
+            window.set_visible(visible);
+        }
     }
 
     pub fn handle_platform_output(